requirements.txt
# Core dependencies
diffusers[torch]==0.26.3
torch

# Required by diffusers
transformers
accelerate
safetensors

# RunPod integration
runpod

# Image processing
pillow

# Network and model downloads
requests
huggingface-hub